firt attempt organizing src
